Tuya Devices Data rename binary_sensor
